adjustments

+++ hello-data-commons/hello-data-nats-spring/src/main/java/ch/bedag/dap/hellodata/commons/nats/annotation/JetStreamSubscribe.java
@@ -39,5 +39,5 @@ public @interface JetStreamSubscribe {
 
     HDEvent event();
 
-    int timeoutMinutes() default 10;
+    long timeoutMinutes() default 10L;
 }
